+++ frontend-react/src/main.tsx
@@ -1,13 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { RouterProvider } from 'react-router'
 import './index.css'
-import App from './app/App.tsx'
+import { router } from './app/router/index.tsx'
 import { AppProviders } from './app/providers/index.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <AppProviders>
-      <App />
+      <RouterProvider router={router} />
     </AppProviders>
   </StrictMode>,
 )